Week12/Day_2/3_financial_modelling/requirements.txt
# Financial Modeling Team Requirements
# Core phidata and AI dependencies
phidata>=2.4.0
openai>=1.0.0

# Data analysis and computation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Financial data
yfinance>=0.2.0

# Visualization and reporting
matplotlib>=3.7.0
plotly>=5.15.0

# Date and time handling
python-dateutil>=2.8.0

# HTTP requests
requests>=2.31.0

# JSON handling
jsonschema>=4.17.0

# Optional: For advanced optimization
cvxpy>=1.3.0

# Optional: For machine learning forecasting
scikit-learn>=1.3.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0